+++ server/fields/idols_field.go
@@ -1,0 +1,26 @@
+package fields
+
+import (
+	"github/tokizuoh/faaaar/server/models"
+
+	"github.com/graphql-go/graphql"
+)
+
+var IdolsFields = &graphql.Field{
+	Type: graphql.NewList(models.IdolType),
+	Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+		ageQuery, ok := p.Args["age"].(int)
+		if ok {
+			result := models.GetSameAgeIdols(models.IdolsByAgeOption{Age: ageQuery})
+			return result, nil
+		} else {
+			result := models.GetSameAgeIdols(models.IdolsByAgeOption{})
+			return result, nil
+		}
+	},
+	Args: graphql.FieldConfigArgument{
+		"age": &graphql.ArgumentConfig{
+			Type: graphql.Int,
+		},
+	},
+}
