+++ server/main.go
@@ -1,9 +1,8 @@
 package main
 
 import (
-	"database/sql"
 	"encoding/json"
-	"fmt"
+	"github/tokizuoh/faaaar/server/fields"
 	"github/tokizuoh/faaaar/server/models"
 	"io/ioutil"
 	"log"
@@ -12,19 +11,6 @@ import (
 	"github.com/graphql-go/graphql"
 	_ "github.com/lib/pq"
 )
-
-type datasourceName struct {
-	host     string
-	port     int
-	user     string
-	password string
-	dbname   string
-	sslmode  string
-}
-
-func getDataSourceNameString(dsn datasourceName) string {
-	return fmt.Sprintf("host=%s port=%d user=%s password=%s dbname=%s sslmode=%s", dsn.host, dsn.port, dsn.user, dsn.password, dsn.dbname, dsn.sslmode)
-}
 
 func readQuery(filepath string) (string, error) {
 	b, err := ioutil.ReadFile(filepath)
@@ -38,57 +24,23 @@ func readQuery(filepath string) (string, error) {
 const QUERY_FILE_PATH = "./query.txt"
 
 func executeQuery(query string) {
-	dsn := datasourceName{
-		host:     "faaaar-db",
-		port:     5432,
-		user:     "postgres",
-		password: "postgres",
-		dbname:   "postgres",
-		sslmode:  "disable",
-	}
-
-	dsnString := getDataSourceNameString(dsn)
-	db, err := sql.Open("postgres", dsnString)
-	defer db.Close()
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
 	scheme, err := graphql.NewSchema(graphql.SchemaConfig{
 		Query: graphql.NewObject(graphql.ObjectConfig{
 			Name: "Query",
 			Fields: graphql.Fields{
-				"idols": &graphql.Field{
-					Type: graphql.NewList(models.IdolType),
-					Resolve: func(p graphql.ResolveParams) (interface{}, error) {
-						ageQuery, ok := p.Args["age"].(int)
-						if ok {
-							result := models.GetSameAgeIdols(db, models.IdolsByAgeOption{Age: ageQuery})
-							return result, nil
-						} else {
-							result := models.GetSameAgeIdols(db, models.IdolsByAgeOption{})
-							return result, nil
-						}
-					},
-					Args: graphql.FieldConfigArgument{
-						"age": &graphql.ArgumentConfig{
-							Type: graphql.Int,
-						},
-					},
-				},
+				"idols": fields.IdolsFields,
 				"units": &graphql.Field{
 					Type: graphql.NewList(models.UnitType),
 					Resolve: func(p graphql.ResolveParams) (interface{}, error) {
 						idolIdQuery, ok := p.Args["idolId"].(int)
 						if ok {
-							result, err := models.GetUnitsByIdolID(db, models.UnitsByIdolIdOption{IdolId: idolIdQuery})
+							result, err := models.GetUnitsByIdolID(models.UnitsByIdolIdOption{IdolId: idolIdQuery})
 							if err != nil {
 								return nil, err
 							}
 							return result, nil
 						} else {
-							result, err := models.GetUnitsByIdolID(db, models.UnitsByIdolIdOption{})
+							result, err := models.GetUnitsByIdolID(models.UnitsByIdolIdOption{})
 							if err != nil {
 								return nil, err
 							}
