+++ server/models/idolsByAge.go
@@ -78,7 +78,39 @@ var IdolType = graphql.NewObject(graphql.ObjectConfig{
 	},
 })
 
-func GetSameAgeIdols(db *sql.DB, o IdolsByAgeOption) []Idol {
+type datasourceName struct {
+	host     string
+	port     int
+	user     string
+	password string
+	dbname   string
+	sslmode  string
+}
+
+func getDataSourceNameString(dsn datasourceName) string {
+	return fmt.Sprintf("host=%s port=%d user=%s password=%s dbname=%s sslmode=%s", dsn.host, dsn.port, dsn.user, dsn.password, dsn.dbname, dsn.sslmode)
+}
+
+func GetSameAgeIdols(o IdolsByAgeOption) []Idol {
+	// TODO: [#30] DB処理共通化 start
+	dsn := datasourceName{
+		host:     "faaaar-db",
+		port:     5432,
+		user:     "postgres",
+		password: "postgres",
+		dbname:   "postgres",
+		sslmode:  "disable",
+	}
+
+	dsnString := getDataSourceNameString(dsn)
+	db, err := sql.Open("postgres", dsnString)
+	defer db.Close()
+
+	if err != nil {
+		log.Fatal(err)
+	}
+	// TODO: [#30] DB処理共通化 end
+
 	stx, err := readSQLFile("./sqls/get_idols_by_age.sql")
 
 	var where string
